disable auth strategy, fix mapping code

+++ helper_scripts/metrics_helper.py
@@ -125,8 +125,9 @@ def cleanup(user_id, agent_id):
     print("cleaning up")
     response = requests.delete(f"{base_url}/users/{user_id}", headers=headers)
     print(response.status_code)
-    response = requests.delete(f"{base_url}/agents/{agent_id}", headers=headers)
-    print(response.status_code)
+    if agent_id:
+        response = requests.delete(f"{base_url}/agents/{agent_id}", headers=headers)
+        print(response.status_code)
 
 
 base_url = "http://localhost:8000/v1"
@@ -155,8 +156,10 @@ print("Setup user info")
 print(response_json)
 
 
+# TODO: make these into tests
 users()
+agent_id = None
 # agent_id = agents()
 # conversation_id = chat(agent_id=agent_id)
 # tools(conversation_id=conversation_id)
-# cleanup(user_id=user_id, agent_id=agent_id)
+cleanup(user_id=user_id, agent_id=agent_id)
